+++ lowcode/icomponent.go
@@ -92,4 +92,10 @@ type IComponent interface {
 	// OnPlayGameWithSet - on playgame with a set
 	OnPlayGameWithSet(gameProp *GameProperty, curpr *sgc7game.PlayResult, gp *GameParams, plugin sgc7plugin.IPlugin,
 		cmd string, param string, ps sgc7game.IPlayerState, stake *sgc7game.Stake, prs []*sgc7game.PlayResult, cd IComponentData, set int) (string, error)
+
+	//----------------------------
+	// Branch
+
+	// GetBranchNum -
+	GetBranchNum() int
 }
